fix(ingest): Fix NULL casting and verify continuous ingestion
- Fixed update_countries function NULL casting for numeric fields
- Verified ingestion daemon running every 15 minutes
- Confirmed 647 signals/cycle flowing correctly
- Total signals: 48,502

+++ backend/app/services/ingest_v2.py
@@ -215,8 +215,8 @@ async def update_countries(pool: asyncpg.Pool):
             SELECT DISTINCT
                 country_code,
                 country_code,
-                NULL,
-                NULL
+                NULL::numeric,
+                NULL::numeric
             FROM signals_v2
             WHERE country_code IS NOT NULL
             AND country_code NOT IN (SELECT code FROM countries_v2)
